refactor: move theme into flags package
Move the Theme into the flags package which better fits what is is.
