+++ core/preferences/src/main/java/com/chesire/nekome/core/preferences/ApplicationPreferences.kt
@@ -5,8 +5,8 @@ import android.content.SharedPreferences
 import androidx.core.content.edit
 import com.chesire.nekome.core.R
 import com.chesire.nekome.core.flags.HomeScreenOptions
+import com.chesire.nekome.core.flags.Theme
 import com.chesire.nekome.core.flags.UserSeriesStatus
-import com.chesire.nekome.core.settings.Theme
 import dagger.hilt.android.qualifiers.ApplicationContext
 import javax.inject.Inject
