+++ features/settings/src/main/java/com/chesire/nekome/app/settings/SettingsFragment.kt
@@ -9,8 +9,8 @@ import androidx.preference.ListPreference
 import androidx.preference.Preference
 import androidx.preference.PreferenceFragmentCompat
 import com.chesire.nekome.core.flags.HomeScreenOptions
+import com.chesire.nekome.core.flags.Theme
 import com.chesire.nekome.core.flags.UserSeriesStatus
-import com.chesire.nekome.core.settings.Theme
 
 /**
  * [SettingsFragment] hosts the configuration options for the application.
