+++ libraries/core/src/test/java/com/chesire/nekome/core/settings/ThemeTests.kt
@@ -2,6 +2,7 @@ package com.chesire.nekome.core.settings
 
 import android.content.Context
 import com.chesire.nekome.core.R
+import com.chesire.nekome.core.flags.Theme
 import io.mockk.every
 import io.mockk.mockk
 import org.junit.Assert.assertEquals
